remove realm, add room classes, refactor
