+++ app/build.gradle
@@ -1,9 +1,6 @@
 apply plugin: 'com.android.application'
-
 apply plugin: 'kotlin-android'
-
 apply plugin: 'kotlin-android-extensions'
-apply plugin: 'realm-android'
 
 def support_version = "27.0.2"
 
@@ -49,14 +46,14 @@ dependencies {
     implementation "android.arch.lifecycle:viewmodel:1.1.0"
     implementation "android.arch.lifecycle:livedata:1.1.0"
     implementation "android.arch.lifecycle:reactivestreams:1.1.0"
-    implementation "android.arch.persistence.room:runtime:1.0.0"
+    implementation "android.arch.persistence.room:runtime:1.1.0-alpha1"
     implementation "android.arch.persistence.room:rxjava2:1.1.0-alpha1"
     implementation "android.arch.paging:runtime:1.0.0-alpha5"
     implementation "android.arch.lifecycle:common-java8:1.1.0"
     testImplementation "junit:junit:4.12"
     kapt "com.android.databinding:compiler:3.0.1"
     kapt "android.arch.lifecycle:compiler:1.1.0"
-    kapt "android.arch.persistence.room:compiler:1.0.0"
+    kapt "android.arch.persistence.room:compiler:1.1.0-alpha1"
 }
 
 kapt {
