+++ app/src/main/java/com/spacelobster/fuckingbills/activities/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.spacelobster.fuckingbills
+package com.spacelobster.fuckingbills.activities
 
 import android.os.Bundle
 import android.support.design.widget.NavigationView
@@ -8,6 +8,7 @@ import android.support.v7.app.ActionBarDrawerToggle
 import android.support.v7.app.AppCompatActivity
 import android.view.Menu
 import android.view.MenuItem
+import com.spacelobster.fuckingbills.R
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.app_bar_main.*
