+++ app/src/main/java/com/spacelobster/fuckingbills/enums/CounterType.kt
@@ -1,8 +1,6 @@
 package com.spacelobster.fuckingbills.enums
 
-/**
- * Created by spacelobster on 8/17/17.
- */
+
 enum class CounterType {
     ELECTRICITY, GAS, WATER
 }
